+++ src/Plutarch/Extra/AssetClass.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE DuplicateRecordFields #-}
+{-# LANGUAGE PolyKinds #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE UndecidableInstances #-}
@@ -34,18 +35,15 @@ module Plutarch.Extra.AssetClass (
   ptoScottEncoding,
   pfromScottEncoding,
   pviaScottEncoding,
-
-  -- * Optics utilities
-  symbolT,
-  nameT,
 ) where
 
 import Data.Aeson (FromJSON, FromJSONKey, ToJSON, ToJSONKey)
 import Data.Tagged (Tagged (Tagged, unTagged), untag)
 import GHC.TypeLits (Symbol)
 import qualified Generics.SOP as SOP
-import Optics.Internal.Optic
-import Optics.Lens (Lens')
+import Optics.Getter (A_Getter, view)
+import Optics.Internal.Optic (A_Lens, Is, (%%))
+import Optics.Label (LabelOptic, LabelOptic', labelOptic)
 import Optics.TH (makeFieldLabelsNoPrefix)
 import Plutarch.Api.V1 (
   PCurrencySymbol,
@@ -162,12 +160,18 @@ passetClassT = phoistAcyclic $
 
 -- | @since 3.10.0
 pconstantClass ::
-  forall (s :: S).
-  AssetClass ->
+  forall (a :: Type) (k :: Type) (s :: S).
+  ( Is k A_Getter
+  , LabelOptic' "symbol" k a CurrencySymbol
+  , LabelOptic' "name" k a TokenName
+  ) =>
+  a ->
   Term s PAssetClass
-pconstantClass (AssetClass sym tk) =
+pconstantClass ac =
   pcon $
-    PAssetClass (pconstantData sym) (pconstantData tk)
+    PAssetClass
+      (pconstantData $ view #symbol ac)
+      (pconstantData $ view #name ac)
 
 -- | @since 3.10.0
 pconstantClassT ::
@@ -204,8 +208,15 @@ psymbolAssetClassT = phoistAcyclic $
 
  @since 3.9.0
 -}
-isAdaClass :: AssetClass -> Bool
-isAdaClass (AssetClass s n) = s == s' && n == n'
+isAdaClass ::
+  forall (a :: Type) (k :: Type).
+  ( Is k A_Getter
+  , LabelOptic' "symbol" k a CurrencySymbol
+  , LabelOptic' "name" k a TokenName
+  ) =>
+  a ->
+  Bool
+isAdaClass ac = view #symbol ac == s' && view #name ac == n'
   where
     (Tagged (AssetClass s' n')) = adaClass
 
@@ -356,7 +367,8 @@ assetClassValue ::
   Tagged unit AssetClass ->
   Tagged unit Integer ->
   Value.Value
-assetClassValue (Tagged (AssetClass sym tk)) q = Value.singleton sym tk $ untag q
+assetClassValue (Tagged (AssetClass sym tk)) q =
+  Value.singleton sym tk $ untag q
 
 ----------------------------------------
 -- Field Labels
@@ -367,14 +379,16 @@ makeFieldLabelsNoPrefix ''PAssetClass
 -- | @since 3.9.0
 makeFieldLabelsNoPrefix ''AssetClass
 
--- | @since 3.10.1
-symbolT ::
-  forall (unit :: Symbol).
-  Lens' (Tagged unit AssetClass) CurrencySymbol
-symbolT = #unTagged %% #symbol
+-- | @since 3.10.2
+instance
+  (k ~ A_Lens, a ~ CurrencySymbol, b ~ CurrencySymbol, tag ~ tag') =>
+  LabelOptic "symbol" k (Tagged tag AssetClass) (Tagged tag' AssetClass) a b
+  where
+  labelOptic = #unTagged %% #symbol
 
--- | @since 3.10.1
-nameT ::
-  forall (unit :: Symbol).
-  Lens' (Tagged unit AssetClass) TokenName
-nameT = #unTagged %% #name
+-- | @since 3.10.2
+instance
+  (k ~ A_Lens, a ~ TokenName, b ~ TokenName, tag ~ tag') =>
+  LabelOptic "name" k (Tagged tag AssetClass) (Tagged tag' AssetClass) a b
+  where
+  labelOptic = #unTagged %% #name